% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CropRaster}
\alias{CropRaster}
\title{CropRaster - Crops (masks) the raster files with the shape file.}
\usage{
CropRaster(filelist = NA, ShapeFile = NA)
}
\arguments{
\item{filelist}{- list of raster files to crop}

\item{ShapeFile}{- mask shapefile used to crop rasters}
}
\description{
Function crops (masks) the raster files with the shape file.
Masked files are stored in the current working directory with original name of file + the sufix provided.
For example, if the input asc files are bio1.asc, bio2.asc, and you give the suffix as 'n', then the output masked files are bio1n.asc, bio2n.asc
If by chance your files does not have any extension, then comment the statement below the note 'For file with extension'.
and uncomment the FileName statement below the note 'For files without extension
}
\examples{
\dontrun{
CropRaster()
}
}

