\name{distancefilter}
\alias{distancefilter}
\title{distancefilter - Raryfy the points in the dataset with threshold distance}
\usage{
  distancefilter(input, threshold, output)
}
\arguments{
  \item{input}{- input file name containing species name,
  Latitude and Longitude}

  \item{threshold}{- value of threshold in (decimal)
  degrees}

  \item{output}{- name of the output file}
}
\description{
  distancefilter - Raryfy the points in the dataset with
  threshold distance
}
\examples{
\dontrun{
distancefilterr()
}
}

