\name{PCARaster}
\alias{PCARaster}
\title{Performs Principle Component Analysis of Raster objects and returns summary and loadings for the componetns.}
\usage{
  PCARaster()
}
\value{
  a summary of the PCA is returnd as a strcture
}
\description{
  Performs Principle Component Analysis of Raster objects
  and returns summary and loadings for the componetns.
}
\examples{
\dontrun{
pcaop = PCARaster()
}
}

